requirements.txt
httpx>=0.27.0
pyswisseph>=2.10.0
pyyaml>=6.0
mlflow>=2.17.0
qdrant-client>=1.12.0
sentence-transformers>=3.0.0
